Refactor report bug URL
