+++ src/api/app/views/webui/package/_show_actions.html.haml
@@ -1,6 +1,6 @@
 - content_for :actions do
   %li.nav-item.action-report-bug
-    = render partial: 'webui/package/show_actions/bugzilla_owner', locals: { url: report_bug_or_bugzilla_url }
+    = render partial: 'webui/package/show_actions/bugzilla_owner', locals: { url: package.report_bug_or_bugzilla_url }
   - if User.session
     - if current_rev && is_working
       = render partial: 'webui/package/show_actions/branch_package', locals: { package: package, project: project,
