+++ src/api/app/views/webui/package/update.js.haml
@@ -13,7 +13,7 @@ resetFormValidation();
                                                               locals: { package: @package, project: @project }))}");
       $('#package-title').html("#{escape_javascript(@package.title.presence || @package.name)}");
       $('.action-report-bug').html("#{escape_javascript(render(partial: 'webui/package/show_actions/bugzilla_owner',
-                                                               locals: { url: @report_bug_or_bugzilla_url }))}");
+                                                               locals: { url: @package.report_bug_or_bugzilla_url }))}");
       setCollapsible();
       $('.in-place-editing,.action-report-bug').animate({ opacity: 1 }, 400, function() {
         $('#flash').html("#{escape_javascript(render(layout: false, partial: 'layouts/webui/flash', object: flash))}");
