+++ src/api/app/views/webui/project/_show_actions.html.haml
@@ -1,6 +1,6 @@
 - content_for :actions do
   %li.nav-item.action-report-bug
-    = render partial: 'webui/project/show_actions/report_bug', locals: { url: report_bug_or_bugzilla_url }
+    = render partial: 'webui/project/show_actions/report_bug', locals: { url: project.report_bug_or_bugzilla_url }
 
   - if project.is_locked? && User.possibly_nobody.can_modify?(project, true)
     = render partial: 'webui/project/show_actions/unlock_project', locals: { project: project }
